Move markTask method from Parser to TaskList
The markTask method pertains more to TaskList as it follows the OOP
principles effectively isolating all behaviour related to a class
within that class.

Let's move markTask from Parser to TaskList to isolate all methods
operating on a TaskList object to TaskList class.

This move allows us to operate solely on TaskList class if there are
any changes to its logic.