Update variable and method names
Names that are not self explanatory makes it harder for the readers.

Let's update names to be more self explanatory to the user, even if
longer names are required, to enhance understandability of code which
is key in team projects.

Longer and more self-explanatory names though longer to type can be
helpful. For example, when someone else debugs your code, they can
read and understand faster.